# api/openapi/api.yaml
openapi: 3.0.0
info:
  title: Schematic REST API
  version: 0.1.0
  description: >-
    This service exposes core functionalities from schematic as REST API
    endpoints

servers:
  - url: /v1

paths:
  /manifest/generate:
    get:
      summary: Endpoint to facilitate manifest generation
      description: Endpoint to create dynamically create metadata manifest files
      parameters:
        - in: query
          name: schema_url
          schema:
            type: string
          description: Data Model URL
          example: >-
            https://raw.githubusercontent.com/Sage-Bionetworks/schematic/develop/tests/data/example.model.jsonld
          required: true
        - in: query
          name: title
          schema:
            type: string
          description: Title of Manifest
          example: Patient Metadata Manifest
          required: false
        - in: query
          name: data_type
          schema:
            type: string
            nullable: true
          description: Data Model Component
          example: Patient
          required: true
        - in: query
          name: oauth
          schema:
            type: boolean
            default: true
          description: OAuth or Service Account
          required: false
        - in: query
          name: use_annotations
          schema:
            type: boolean
            default: false
          description: To Use Annotations
          required: false
        - in: query
          name: dataset_id
          schema:
            type: string
            nullable: true
          description: Dataset SynID
          required: true
      operationId: api.routes.get_manifest_route
      responses:
        "201":
          description: Googlesheet link created
          content:
            application/json:
              schema:
                type: string
      tags:
        - Manifest Operations
  /model/validate:
    post:
      summary: Endpoint to facilitate manifest validation
      description: Endpoint to validate metadata manifest files
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                # csv_file will be the field name in
                # this multipart request
                csv_file:
                  type: string
                  format: binary
      parameters:
        - in: query
          name: schema_url
          schema:
            type: string
          description: Data Model URL
          example: >-
            https://raw.githubusercontent.com/Sage-Bionetworks/schematic/develop/tests/data/example.model.jsonld
          required: true
        - in: query
          name: data_type
          schema:
            type: string
            nullable: true
          description: Data Model Component
          example: Patient
          required: true
      operationId: api.routes.validate_manifest_route
      responses:
        "200":
          description: Manifest Validated
          content:
            application/json:
              schema:
                type: array
                items:
                  type: array
                  items:
                    anyOf:
                      - type: integer
                      - type: string
                      - type: array
                        items:
                          type: string
      tags:
        - Model Operations
  /model/submit:
    post:
      summary: Endpoint to facilitate manifest submission
      description: Endpoint to submit annotated manifest files
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                csv_file:
                  type: string
                  format: binary
      parameters:
        - in: query
          name: schema_url
          schema:
            type: string
          description: Data Model URL
          example: >-
            https://raw.githubusercontent.com/Sage-Bionetworks/schematic/develop/tests/data/example.model.jsonld
          required: true
        - in: query
          name: data_type
          schema:
            type: string
            nullable: true
          description: Data Model Component
          example: Patient
          required: true
        - in: query
          name: dataset_id
          schema:
            type: string
            nullable: true
          description: Dataset SynID
          required: true
        - in: query
          name: input_token
          schema:
            type: string
            nullable: true
          description: Token
          required: true
      operationId: api.routes.submit_manifest_route
      responses:
        "200":
          description: Manifest submitted
          content:
            application/json:
              schema:
                type: boolean
        "500":
          description: Check schematic log
      tags:
        - Model Operations